Organize source files into dedicated folders
