+++ src/gameobjects/Symbol.ts
